Problem: duck-typing in Zig
When you have a function taking `type` or `anytype` parameters, it's not
trivial to tell what that type should be like.

One can write good documentation but that requires extra maintenance efforts
and is not very formalized.

There's `std.meta.trait` but it is limited in terms of what it can do, and particulary,
it's not great at identifying the cause of contract (trait) and have limited composition.

Solution: formalize contracts with causation

This is just a first check-in, it doesn't do much in terms of library completeness
(in fact, there are only three primitives), but it (hopefully) provides a good
enough foundation.