Problem: impossible to test against a type of a type
Solution: implement `isType`

This change illuminated a necessity to add `.reason` field to
`Invalid` to ease the debugging. Well, to be entirely honest,
this specific contract might not have needed, but I think that
this is a good one to keep. I think more complex contracts will
make use of it.