Problem: complex contract checks
Let's imagine that we want to check if there's a declaration,
and if only there's one, check if it matches certain contract.

Solution: add `andThen` and `orThen` helpers

They will verify contracts only if the given contract have succeeded
or failed, respectively.

As a part of the implementation, `hasDecl` contract has been added
to enable contract validation in `andThen` and `orElse`.