Problem: git complaining about .gitattributes:2
Solution: remove the comment

gitattributes documentation says this:

```
A gitattributes file is a simple text file that gives attributes to pathnames.

Each line in gitattributes file is of form:

pattern attr1 attr2 ...

That is, a pattern followed by an attributes list, separated by whitespaces.
Leading and trailing whitespaces are ignored. Lines that begin with # are
ignored. Patterns that begin with a double quote are quoted in C style. When
the pattern matches the path in question, the attributes listed on the line are
given to the path.
```

My reading is that it's only lines that *begin* with # are comments