Problem: limited extensibility
Composition of contracts is great, but what if we want to establish
further conceptual contracts over data types?

Solution: experiment with `contracts` declarations
that would allow types to supply information to actual contracts.

First example is the establishment of equivalency -- types that are not
equal but "equivalent" (whatever that means for the domain logic).

This example necessitates addition of `contracts.isEquivalent(type) -> bool`
function to at least one of the types.

I ran into the case of not being able to compare generic function types, so
I hand-rolled one, but it is still imperfect as it can't get type information
of the return (or so it seems).